requirements-conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiohttp=3.8.5=py311h2725bcf_0
aiosignal=1.3.1=pyhd8ed1ab_0
async-timeout=4.0.3=pyhd8ed1ab_0
attrs=23.1.0=pyh71513ae_1
blas=1.0=mkl
blinker=1.6.2=pyhd8ed1ab_0
bottleneck=1.3.5=py311hb9e55a9_0
brotli=1.1.0=h0dc2134_0
brotli-bin=1.1.0=h0dc2134_0
brotli-python=1.1.0=py311hdf8f085_0
bzip2=1.0.8=h0d85af4_4
ca-certificates=2023.7.22=h8857fd0_0
cachetools=4.2.4=pyhd8ed1ab_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py311ha86e640_3
charset-normalizer=3.2.0=pyhd8ed1ab_0
click=8.1.7=unix_pyh707e725_0
configargparse=1.7=pyhd8ed1ab_0
contourpy=1.1.0=py311h5fe6e05_0
cryptography=41.0.3=py311h892b619_0
cycler=0.11.0=pyhd8ed1ab_0
daal4py=2023.1.1=py311h85bffb1_0
dal=2023.1.1=ha357a0b_48680
docopt=0.6.2=py_1
filelock=3.9.0=py311hecd8cb5_0
fonttools=4.42.1=py311h2725bcf_0
freetype=2.12.1=h3f81eb7_1
frozenlist=1.4.0=py311h2725bcf_0
gmp=6.2.1=he9d5cce_3
gmpy2=2.1.2=py311h1c2e9e1_0
google-auth=1.35.0=pyh6c4a22f_0
google-auth-oauthlib=1.0.0=pyhd8ed1ab_0
gspread=5.7.1=pyh1a96a4e_0
httplib2=0.22.0=pyhd8ed1ab_0
idna=3.4=pyhd8ed1ab_0
intel-openmp=2023.1.0=ha357a0b_43547
jinja2=3.1.2=py311hecd8cb5_0
joblib=1.2.0=py311hecd8cb5_0
kiwisolver=1.4.5=py311h5fe6e05_0
lcms2=2.15=h2dcdeff_1
lerc=4.0.0=hb486fe8_0
libbrotlicommon=1.1.0=h0dc2134_0
libbrotlidec=1.1.0=h0dc2134_0
libbrotlienc=1.1.0=h0dc2134_0
libcxx=16.0.6=hd57cbcb_0
libdeflate=1.18=hac1461d_0
libexpat=2.5.0=hf0c8a7f_1
libffi=3.4.2=h0d85af4_5
libgfortran=5.0.0=11_3_0_hecd8cb5_28
libgfortran5=11.3.0=h9dfd629_28
libjpeg-turbo=2.1.5.1=hb7f2c08_0
libopenblas=0.3.21=h54e7dc3_0
libpng=1.6.39=ha978bb4_0
libsqlite=3.43.0=h58db7d2_0
libtiff=4.5.1=hf955e92_1
libwebp-base=1.3.1=h0dc2134_0
libxcb=1.15=hb7f2c08_0
libzlib=1.2.13=h8a1eda9_5
llvm-openmp=14.0.6=h0dcd299_0
markupsafe=2.1.1=py311h6c40b1e_0
matplotlib=3.7.1=py311hecd8cb5_1
matplotlib-base=3.7.1=py311h2bf763f_0
mkl=2023.1.0=h8e150cf_43559
mkl-service=2.4.0=py311h6c40b1e_1
mkl_fft=1.3.6=py311hdb55bb0_1
mkl_random=1.2.2=py311hdb55bb0_1
mpc=1.1.0=h6ef4df4_1
mpfr=4.0.2=h9066e36_1
mpi=1.0=mpich
mpich=4.1.1=hc226d05_0
mpmath=1.3.0=py311hecd8cb5_0
multidict=6.0.4=py311h5547dcb_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.4=hf0c8a7f_0
networkx=3.1=py311hecd8cb5_0
numexpr=2.8.4=py311h728a8a3_1
numpy=1.25.2=py311h728a8a3_0
numpy-base=1.25.2=py311h53bf9ac_0
oauth2client=4.1.3=py_0
oauthlib=3.2.2=pyhd8ed1ab_0
openjpeg=2.5.0=h13ac156_2
openssl=3.1.2=h8a1eda9_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.3=py311hdb55bb0_0
pillow=10.0.0=py311h7cb0e2d_0
pip=23.2.1=pyhd8ed1ab_0
pipreqs=0.4.13=pyhd8ed1ab_0
pthread-stubs=0.4=hc929b4f_1001
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pyjwt=2.8.0=pyhd8ed1ab_0
pyopenssl=23.2.0=pyhd8ed1ab_1
pyparsing=3.1.1=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.11.5=h30d4d87_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.11=3_cp311
pytorch=2.0.1=py3.11_0
pytz=2021.3=pyhd3eb1b0_0
pyu2f=0.1.5=pyhd8ed1ab_0
readline=8.2=h9e318b2_1
requests=2.31.0=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scikit-learn=1.2.2=py311hcec6c5f_0
scikit-learn-intelex=2023.1.1=py311hecd8cb5_0
scipy=1.11.1=py311h224febf_0
setuptools=68.1.2=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sympy=1.11.1=py311hecd8cb5_0
tbb=2021.8.0=ha357a0b_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h5dbffcc_0
tornado=6.3.3=py311h2725bcf_0
typing-extensions=4.7.1=hd8ed1ab_0
typing_extensions=4.7.1=pyha770c72_0
tzdata=2023c=h71feb2d_0
urllib3=2.0.4=pyhd8ed1ab_0
wheel=0.41.2=pyhd8ed1ab_0
xorg-libxau=1.0.11=h0dc2134_0
xorg-libxdmcp=1.1.3=h35c211d_0
xz=5.2.6=h775f41a_0
yarg=0.1.9=py_1
yarl=1.9.2=py311h2725bcf_0
zstd=1.5.5=h829000d_0
